Preserve server directory of source code
This mod re-organizes all code under the server directory once more to
provide a more clean abstraction to the overall source.
