+++ Dockerfile
@@ -25,8 +25,8 @@
 ARG OPENAPI_IMAGE=openapitools/openapi-generator-cli:v4.1.2
 ARG ALPINE_BASE_IMAGE=artifactory.algol60.net/docker.io/alpine:3.14
 
-# Generate API
-FROM $OPENAPI_IMAGE as codegen_api
+# Generate Code
+FROM $OPENAPI_IMAGE as codegen
 WORKDIR /app
 COPY api/openapi.yaml api/openapi.yaml
 COPY config/autogen-server.json config/autogen-server.json
@@ -36,34 +36,26 @@ RUN /usr/local/bin/docker-entrypoint.sh generate \
     -o lib \
     -c config/autogen-server.json \
     --generate-alias-as-model
-
-# Add autogenerated client code for V2 Agents
-#FROM $OPENAPI_IMAGE as codegen_client
-#WORKDIR /app
-#COPY api/openapi.yaml api/openapi.yaml
-#COPY config/autogen-server.json config/autogen-server.json
-#COPY src/server/.openapi-generator-ignore lib/server/.openapi-generator-ignore
-#RUN /usr/local/bin/docker-entrypoint.sh generate \
-#    -i api/openapi.yaml \
-#    -g python \
-#    -o lib \
-#    -c config/autogen-server.json
+RUN /usr/local/bin/docker-entrypoint.sh generate \
+    -i api/openapi.yaml \
+    -g python \
+    -o lib/bos/client \
+    -c config/autogen-client.json
 
 # Base image
 FROM $ALPINE_BASE_IMAGE as base
 WORKDIR /app
 # We apply all generated code first
-#COPY --from=codegen_cli /app .
-COPY --from=codegen_api /app .
+COPY --from=codegen /app/lib/ /app/lib
 # Move autogenerated server requirements aside so that they can be referenced by
 # project-wide requirements.txt
-RUN mv lib/requirements.txt lib/requirements_server.txt
+RUN mv lib/requirements.txt lib/bos/server/requirements.txt
 # Then copy all src into the base image
 COPY src/bos/ /app/lib/bos
 COPY constraints.txt requirements.txt .
 # Update packages to avoid security problems
 RUN apk add --upgrade --no-cache apk-tools busybox && \
-	apk update && \
+    apk update && \
     apk add --no-cache gcc g++ python3-dev py3-pip musl-dev libffi-dev openssl-dev && \
     apk -U upgrade --no-cache && \
     pip3 install --no-cache-dir -U pip && \
