+++ config/uwsgi.ini
@@ -6,7 +6,7 @@ master=true
 vacuum=true
 die-on-term=true
 app=bos
-module=bos.__main__
+module=bos.server.__main__
 callable=app
 processes=8
 threads=16
