+++ requirements.txt
@@ -3,7 +3,8 @@
 --index-url https://arti.dev.cray.com:443/artifactory/api/pypi/pypi-remote/simple
 --extra-index-url http://artifactory.algol60.net/artifactory/csm-python-modules/simple
 -c constraints.txt
--r lib/requirements_server.txt
+-r lib/bos/server/requirements.txt
+-r lib/bos/client/requirements.txt
 
 kubernetes
 redis
@@ -17,5 +18,6 @@ liveness
 # The purpose of this file is to contain python runtime requirements
 # for controller code, e.g., code authored by developers, as opposed to
 # automatically-generated code. Requirements from automatically-generated
-# code are stored in lib/server/requirements.txt (within the associated
-# docker image), and are referenced above.
+# code are stored in lib/bos/server/requirements.txt (within the associated
+# docker image), and are referenced above. Similarly, client specific code
+# that is automatically generated during build time is in the client submodule.
