+++ src/bos/server/test/test_version_controller.py
@@ -1,0 +1,67 @@
+# coding: utf-8
+# Copyright 2019, 2021 Hewlett Packard Enterprise Development LP
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included
+# in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL
+# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+# OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+# ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+# OTHER DEALINGS IN THE SOFTWARE.
+#
+# (MIT License)
+
+from __future__ import absolute_import
+
+from flask import json
+from six import BytesIO
+
+from bos.server.models.problem_details import ProblemDetails  # noqa: E501
+from bos.server.models.version import Version  # noqa: E501
+from bos.server.test import BaseTestCase
+from bos.server.controllers.base import root_get
+
+from nose.tools import nottest
+
+
+class TestVersionController(BaseTestCase):
+    """VersionController integration test stubs"""
+
+    @nottest
+    def test_get_version(self):
+        """Test case for get_version
+
+        API version
+        """
+        response = self.client.open(
+            '/apis/bos/v1',
+            method='GET')
+        self.assert200(response,
+                       'Response body is : ' + response.data.decode('utf-8'))
+
+    @nottest
+    def test_get_versions(self):
+        """Test case for get_versions
+
+        API versions
+        """
+        response = self.client.open(
+            '/apis/bos/',
+            method='GET')
+        self.assert200(response,
+                       'Response body is : ' + response.data.decode('utf-8'))
+
+
+if __name__ == '__main__':
+    import unittest
+    unittest.main()
