+++ src/bos/server/v1_v2_migration.py
@@ -24,11 +24,11 @@ import json
 import logging
 import requests
 
-from server.bos.dbclient import BosEtcdClient
-from bos.operators.utils import requests_retry_session
-import server.bos.redis_db_utils as dbutils
+from bos.server.dbclient import BosEtcdClient
+from bos.server.operators.utils import requests_retry_session
+import bos.server.redis_db_utils as dbutils
 
-LOGGER = logging.getLogger('bos.v1_v2_migration')
+LOGGER = logging.getLogger('bos.server.v1_v2_migration')
 DB = dbutils.get_wrapper(db='session_templates')
 BASEKEY = "/sessionTemplate"
